Implement fetching and displaying of profile data
The commit includes the implementation of fetching and displaying user profile information in the profile info section. User profile data is fetched by updating the profile reducer with a new setProfile action, which sets the retrieved profile data into the app state. To aid in application functionality, new utility functions are added. Moreover, user experience has been improved by resetting the input field after adding a new item and an automatic redirection to the profile page on the first render.